fix: tg id is BigInteger

+++ app/company/tables.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Table
+from sqlalchemy import Column, Integer, String, Table, BigInteger
 
 from app.core.database import metadata
 
@@ -8,5 +8,5 @@ company_table = Table(
     Column('id', Integer, primary_key=True),
     Column('name', String, nullable=False),
     Column('code', String(3), unique=True, nullable=False),
-    Column('owner_tg_id', Integer, nullable=False),
+    Column('owner_tg_id', BigInteger, nullable=False),
 )
