+++ app/employee/tables.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Boolean, Column, DateTime, Float, ForeignKey, Integer, String, Table
+from sqlalchemy import Boolean, Column, DateTime, Float, ForeignKey, Integer, String, Table, BigInteger
 
 from app.core.database import metadata
 
@@ -6,7 +6,7 @@ employee_table = Table(
     'employee',
     metadata,
     Column('id', Integer, primary_key=True),
-    Column('telegram_id', Integer, nullable=False),
+    Column('telegram_id', BigInteger, nullable=False),
     Column('company_id', Integer, ForeignKey('company.id', ondelete='CASCADE'), nullable=False),
     Column('is_active', Boolean, nullable=False),
     Column('is_admin', Boolean, nullable=False),
